File renaming and small changes
